Moved _invert_continua to interface
